Rename folder from model to entity
